fix go run for integration
