+++ cmd/integration/commands/flags.go
@@ -1,4 +1,4 @@
-package main
+package commands
 
 import "github.com/spf13/cobra"
