+++ cmd/integration/main.go
@@ -1,56 +1,8 @@
 package main
 
-import (
-	"context"
-	"fmt"
-	"os"
-	"os/signal"
-	"syscall"
-
-	"github.com/ledgerwatch/turbo-geth/cmd/utils"
-	"github.com/ledgerwatch/turbo-geth/internal/debug"
-	"github.com/ledgerwatch/turbo-geth/log"
-	"github.com/spf13/cobra"
-)
-
-var rootCmd = &cobra.Command{
-	Use:   "integration",
-	Short: "long and heavy integration tests for turbo-geth",
-	PersistentPreRun: func(cmd *cobra.Command, args []string) {
-		if err := debug.SetupCobra(cmd); err != nil {
-			panic(err)
-		}
-	},
-	PersistentPostRun: func(cmd *cobra.Command, args []string) {
-		debug.Exit()
-	},
-}
-
-func init() {
-	utils.CobraFlags(rootCmd, append(debug.Flags, utils.MetricsEnabledFlag, utils.MetricsEnabledExpensiveFlag))
-}
+import "github.com/ledgerwatch/turbo-geth/cmd/integration/commands"
 
 func main() {
-	if err := rootCmd.ExecuteContext(rootContext()); err != nil {
-		fmt.Println(err)
-		os.Exit(1)
-	}
+	commands.Execute()
 }
 
-func rootContext() context.Context {
-	ctx, cancel := context.WithCancel(context.Background())
-	go func() {
-		ch := make(chan os.Signal, 1)
-		signal.Notify(ch, os.Interrupt, syscall.SIGTERM)
-		defer signal.Stop(ch)
-
-		select {
-		case <-ch:
-			log.Info("Got interrupt, shutting down...")
-		case <-ctx.Done():
-		}
-
-		cancel()
-	}()
-	return ctx
-}
